Restructuring: Move validataclass decorator and validataclass_field function to separate modules
