+++ src/validataclass/dataclasses/__init__.py
@@ -4,6 +4,7 @@ Copyright (c) 2022, binary butterfly GmbH and contributors
 Use of this source code is governed by an MIT-style license that can be found in the LICENSE file.
 """
 
-from .dataclass_defaults import Default, DefaultFactory, DefaultUnset, NoDefault
-from .dataclass_mixins import ValidataclassMixin
-from .dataclasses import validataclass, validataclass_field
+from .defaults import Default, DefaultFactory, DefaultUnset, NoDefault
+from .validataclass import validataclass
+from .validataclass_field import validataclass_field
+from .validataclass_mixin import ValidataclassMixin
